src/generator.py
import torch
import torch.nn as nn
import torchvision.transforms as transforms
from PIL import Image
import numpy as np

class ImageGenerator:
    def __init__(self):
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.transform = transforms.Compose([
            transforms.Resize((256, 256)),
            transforms.ToTensor(),
            transforms.Normalize(mean=[0.5, 0.5, 0.5], std=[0.5, 0.5, 0.5])
        ])
    
    def load_image(self, image_path):
        image = Image.open(image_path).convert('RGB')
        return self.transform(image).unsqueeze(0)
    
    def preprocess_image(self, image):
        # Add preprocessing logic here
        return image
    
    def generate(self, input_image):
        # Add generation logic here
        return input_image
    
    def save_image(self, tensor, output_path):
        # Denormalize and convert to PIL Image
        image = tensor.cpu().squeeze(0)
        image = image * 0.5 + 0.5
        image = transforms.ToPILImage()(image)
        image.save(output_path)