requirements.txt
torch>=1.7.0
torchvision>=0.8.1
pillow>=7.0.0
numpy>=1.19.2
tqdm>=4.51.0
matplotlib>=3.3.2